pyproject.toml
[project]
name = "bowling-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.8",
]
